Explicitly cast enum to int inside CHECK()
Somehow using enum results in wrong comparison being done when using gcc
5.3 (the test run when using this compiler fails because the RHS value
is "true" and not the numeric value of the enum element), so add a cast
to make this work.

+++ tests/menu/menu.cpp
@@ -622,7 +622,7 @@ void MenuTestCase::Events()
     wxYield();
 
     const wxCommandEvent& ev2 = handler.GetEvent();
-    CHECK( ev2.GetId() == MenuTestCase_SelectAll );
+    CHECK( ev2.GetId() == static_cast<int>(MenuTestCase_SelectAll) );
 
     // Now create a text control which uses the same accelerator for itself and
     // check that when the text control has focus, the accelerator does _not_
